bugfix: debouncer causes i2c to hang
was an unsigned byte overflow issue when refresh = 1

+++ keyboard-specific-builds/diverge-tm-2/diverge-tm-2-slave/animus/mod_tapDance.ino
@@ -25,7 +25,7 @@ const byte modMethod(MAX_INDEX) = 20;
 const byte modMethod(MAX_KEYS) = 3;
 const byte modMethod(PADDING) = 2;
 
-byte modMethod(PressTimer)[modMethod(MAX_INDEX)];
+uint16_t modMethod(PressTimer)[modMethod(MAX_INDEX)];
 byte modMethod(PressCount)[modMethod(MAX_INDEX)];
 boolean modMethod(KeyIsDown)[modMethod(MAX_INDEX)];
